+++ src/compas_timber/connections/butt_joint.py
@@ -22,16 +22,10 @@ class ButtJoint(Joint):
         The cross beam to be joined.
     mill_depth : float
         The depth of the pocket to be milled in the cross beam. This will be ignored if `butt_plane` is provided.
-    small_beam_butts : bool, default False
-        If True, the beam with the smaller cross-section will be trimmed. Otherwise, the main beam will be trimmed.
     modify_cross : bool, default False
         If True, the cross beam will be extended to the opposite face of the main beam and cut with the same plane.
     butt_plane : :class:`~compas.geometry.Plane`, optional
         The plane used to cut the main beam. If not provided, the closest side of the cross beam will be used.
-    back_plane : :class:`~compas.geometry.Plane`, optional
-        The plane used to cut the cross beam. If not provided, the back side of the main beam will be used.
-    reject_i : bool, default False
-        If True, the joint will reject beams in I topology.
 
     Attributes
     ----------
@@ -41,12 +35,9 @@ class ButtJoint(Joint):
         The cross beam to be joined.
     mill_depth : float
         The depth of the pocket to be milled in the cross beam.
-    small_beam_butts : bool, default False
-        If True, the beam with the smaller cross-section will be trimmed. Otherwise, the main beam will be trimmed.
     modify_cross : bool, default False
         If True, the cross beam will be extended to the opposite face of the main beam and cut with the same plane.
-    reject_i : bool, default False
-        If True, the joint will reject beams in I topology.
+
 
     """
 
@@ -58,29 +49,22 @@ class ButtJoint(Joint):
         data["main_beam_guid"] = self.main_beam_guid
         data["cross_beam_guid"] = self.cross_beam_guid
         data["mill_depth"] = self.mill_depth
-        data["small_beam_butts"] = self.small_beam_butts
         data["modify_cross"] = self.modify_cross
-        data["reject_i"] = self.reject_i
         data["butt_plane"] = self.butt_plane
-        data["back_plane"] = self.back_plane
         return data
 
-    def __init__(self, main_beam=None, cross_beam=None, mill_depth=None, small_beam_butts=False, modify_cross=True, reject_i=False, butt_plane=None, back_plane=None, **kwargs):
+    def __init__(self, main_beam=None, cross_beam=None, mill_depth=None, modify_cross=True, butt_plane=None, **kwargs):
         super(ButtJoint, self).__init__(**kwargs)
         self.main_beam = main_beam
         self.cross_beam = cross_beam
         self.main_beam_guid = kwargs.get("main_beam_guid", None) or str(main_beam.guid)
         self.cross_beam_guid = kwargs.get("cross_beam_guid", None) or str(cross_beam.guid)
         self.mill_depth = mill_depth
-        self.small_beam_butts = small_beam_butts
         self.modify_cross = modify_cross
-        self.reject_i = reject_i
         self.butt_plane = butt_plane
-        self.back_plane = back_plane
         self.features = []
 
         # update the main and cross beams based on the joint parameters
-        self.update_beam_roles()
 
     @property
     def elements(self):
@@ -104,15 +88,6 @@ class ButtJoint(Joint):
         if self.reject_i and ref_side_index in [4, 5]:
             raise BeamJoiningError(beams=self.elements, joint=self, debug_info="Beams are in I topology and reject_i flag is True")
         return ref_side_index
-
-    def update_beam_roles(self):
-        """Flips the main and cross beams based on the joint parameters.
-        Prioritizes the beam with the smaller cross-section if `small_beam_butts` is True.
-
-        """
-        if self.small_beam_butts:
-            if self.main_beam.width * self.main_beam.height > self.cross_beam.width * self.cross_beam.height:
-                self.main_beam, self.cross_beam = self.cross_beam, self.main_beam
 
     def add_extensions(self):
         """Calculates and adds the necessary extensions to the beams.