+++ CHANGELOG.md
@@ -31,6 +31,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added tasks `update-gh-header` to update the version in the header of the GH components.
 * Added new `compas_timber.connections.XNotchJoint`.
 * Added a proxy class for `Pocket` BTLx processing for performance optimization. 
+* Added `topology` to class `Joint`.
+* Added `location` to class `Joint`.
+* Added `NBeamKDTreeAnalyzer` to `compas_timber.connections`.
+* Added `TripletAnalyzer` to `compas_timber.connections`.
+* Added `QuadAnalyzer` to `compas_timber.connections`.
+* Added `CompositeAnalyzer` to `compas_timber.connections`.
+* Added method `connect_adjacent_beams` to `TimberModel`.
 * Added `PlateJoint`.
 * Added `PlateButtJoint`.
 * Added `PlateMiterJoint`.
@@ -46,11 +53,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Processings which are not the result of joinery are now serialized with `TimberElement`.
 * Fixed visualization bug in `Plate` due to loft resulting in flipped volume.
 * Fixed a few bugs in the `WallPopulator` workflow including GH component updates.
+* Renamed `NullJoint` to `GenericJoint`.
 * Fixed bug in show_ref_faces GH component.
+* `BTLxProcessing.ref_side_index` defaults to `0` if not set, instead of the invalid `None`.
 
 ### Removed
 
 * Removed Grasshopper after-install plugin. Components should be installed via Rhino's Plugin Manager.
+* Removed `get_face_most_towards_beam` from `Joint` as not used anywhere.
+* Removed `get_face_most_ortho_to_beam` from `Joint` as not used anywhere.
 
 
 ## [0.16.2] 2025-05-07
@@ -62,6 +73,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fixed max recursion depth error when copying `TimberModel`/`Beam` with proxy processings.
 
 ### Removed
+
 
 
 ## [0.16.1] 2025-04-30